terminal-emulator: 实现 Rust 颜色管理、标题栈和行绘图功能
新增功能:
- TerminalColors 259 色管理 (OSC 4/10/11/12/104/110/111/112)
- 标题栈支持 (OSC 22/23)
- 行绘图字符集 G0/G1 切换 (ESC (0 / )0 / B 和 SO/SI)
- 滚动计数器和自动滚动禁用
- SGR 58/59 下划线颜色支持
- 保存/恢复光标增强 (包含行绘图状态和颜色)
- RIS 完整重置功能

测试:
- 新增 13 个一致性测试用例
- 所有 90 个测试通过

技术细节:
- 修复位定义确保与 Java TextStyle 兼容
- 实现 OSC 序列参数解析
- 添加 DEFAULT_COLORSCHEME 常量导出
- 优化 execute_control 处理 SO/SI 切换

+++ terminal-emulator/src/main/rust/src/lib.rs
@@ -95,7 +95,7 @@ pub unsafe extern "system" fn Java_com_termux_terminal_TerminalEmulator_mapLineD
 
         let dest_vec: Vec<u16> = src_vec
             .par_iter()
-            .map(|&b| utils::map_line_drawing(b as u8))
+            .map(|&b| utils::map_line_drawing(b as u8) as u16)
             .collect();
 
         let _ = env.set_char_array_region(jni::objects::JCharArray::from_raw(dest), 0, &dest_vec);
